rename vat-countries to contributions-only-countries
