+++ .github/workflows/ci-typescript.yml
@@ -71,7 +71,7 @@ jobs:
           - newsletter-acquisition
           - multiple-account-api
           - observer-benefits-api
-          - vat-countries-list-api
+          - contributions-only-countries-api
           # MARKER new-lambda: github-action
     runs-on: ubuntu-latest
     permissions:
