+++ BUILDCHECK.md
@@ -135,11 +135,11 @@ For further details, see [buildcheck/README.md](./buildcheck/README.md)
 - [handlers/observer-benefits-api/riff-raff.yaml](handlers/observer-benefits-api/riff-raff.yaml)
 - [handlers/observer-benefits-api/tsconfig.json](handlers/observer-benefits-api/tsconfig.json)
 - [handlers/observer-benefits-api/BUILDCHECK.md](handlers/observer-benefits-api/BUILDCHECK.md)
-- [handlers/vat-countries-list-api/jest.config.js](handlers/vat-countries-list-api/jest.config.js)
-- [handlers/vat-countries-list-api/package.json](handlers/vat-countries-list-api/package.json)
-- [handlers/vat-countries-list-api/riff-raff.yaml](handlers/vat-countries-list-api/riff-raff.yaml)
-- [handlers/vat-countries-list-api/tsconfig.json](handlers/vat-countries-list-api/tsconfig.json)
-- [handlers/vat-countries-list-api/BUILDCHECK.md](handlers/vat-countries-list-api/BUILDCHECK.md)
+- [handlers/contributions-only-countries-api/jest.config.js](handlers/contributions-only-countries-api/jest.config.js)
+- [handlers/contributions-only-countries-api/package.json](handlers/contributions-only-countries-api/package.json)
+- [handlers/contributions-only-countries-api/riff-raff.yaml](handlers/contributions-only-countries-api/riff-raff.yaml)
+- [handlers/contributions-only-countries-api/tsconfig.json](handlers/contributions-only-countries-api/tsconfig.json)
+- [handlers/contributions-only-countries-api/BUILDCHECK.md](handlers/contributions-only-countries-api/BUILDCHECK.md)
 - [modules/aws/jest.config.js](modules/aws/jest.config.js)
 - [modules/aws/package.json](modules/aws/package.json)
 - [modules/aws/tsconfig.json](modules/aws/tsconfig.json)
@@ -148,6 +148,10 @@ For further details, see [buildcheck/README.md](./buildcheck/README.md)
 - [modules/bigquery/package.json](modules/bigquery/package.json)
 - [modules/bigquery/tsconfig.json](modules/bigquery/tsconfig.json)
 - [modules/bigquery/BUILDCHECK.md](modules/bigquery/BUILDCHECK.md)
+- [modules/contributions-only-countries-list/jest.config.js](modules/contributions-only-countries-list/jest.config.js)
+- [modules/contributions-only-countries-list/package.json](modules/contributions-only-countries-list/package.json)
+- [modules/contributions-only-countries-list/tsconfig.json](modules/contributions-only-countries-list/tsconfig.json)
+- [modules/contributions-only-countries-list/BUILDCHECK.md](modules/contributions-only-countries-list/BUILDCHECK.md)
 - [modules/email/jest.config.js](modules/email/jest.config.js)
 - [modules/email/package.json](modules/email/package.json)
 - [modules/email/tsconfig.json](modules/email/tsconfig.json)
