+++ cdk/lib/contributions-only-countries-api.test.ts
@@ -1,12 +1,12 @@
 import { App } from 'aws-cdk-lib';
 import { Template } from 'aws-cdk-lib/assertions';
-import { VatCountriesListApi } from './vat-countries-list-api';
+import { ContributionsOnlyCountriesApi } from './contributions-only-countries-api';
 
-describe('The Vat countries list api stack', () => {
+describe('The contributions-only countries list api stack', () => {
 	it('matches the snapshot', () => {
 		const app = new App();
-		const codeStack = new VatCountriesListApi(app, 'CODE');
-		const prodStack = new VatCountriesListApi(app, 'PROD');
+		const codeStack = new ContributionsOnlyCountriesApi(app, 'CODE');
+		const prodStack = new ContributionsOnlyCountriesApi(app, 'PROD');
 
 		expect(Template.fromStack(codeStack).toJSON()).toMatchSnapshot();
 		expect(Template.fromStack(prodStack).toJSON()).toMatchSnapshot();
