+++ cdk/lib/contributions-only-countries-api.ts
@@ -3,17 +3,18 @@ import { SrApiLambda } from './cdk/SrApiLambda';
 import type { SrStageNames } from './cdk/SrStack';
 import { SrStack } from './cdk/SrStack';
 
-export class VatCountriesListApi extends SrStack {
+export class ContributionsOnlyCountriesApi extends SrStack {
 	constructor(scope: App, stage: SrStageNames) {
-		super(scope, { stage, app: 'vat-countries-list-api' });
+		super(scope, { stage, app: 'contributions-only-countries-api' });
 
 		new SrApiLambda(this, 'Lambda', {
 			lambdaOverrides: {
-				description: 'A lambda that returns a list of VAT compliant countries',
+				description:
+					'A lambda that returns a list of countries that are contributions-only',
 			},
 			monitoring: {
 				errorImpact:
-					'users from VAT compliant countries will see the choice cards in epics/banners and https://support.theguardian.com/contribute screen',
+					'users from contributions-only countries will see the choice cards in epics/banners and support site',
 			},
 
 			isPublic: true, // Don't create an API Key for this lambda
