+++ handlers/contributions-only-countries-api/README.md
@@ -1,0 +1,12 @@
+# VAT countries list API
+
+This API is an API Gateway triggered lambda to return the list of
+VAT compliant countries. It is intended to be used by SDC project
+to prevent users from the countries from the list to see any products besides contributions. The components that should be hidden are: choice cards in epics and banners and in support site.
+
+It is a public API endpoint
+
+# Urls
+
+- PROD: https://contributions-only-countries-api.support.guardianapis.com/
+- CODE: https://contributions-only-countries-api.support.guardianapis.com/
