+++ handlers/contributions-only-countries-api/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "contributions-only-countries-api",
+  "scripts": {
+    "test": "jest --group=-integration",
+    "it-test": "jest --group=integration",
+    "type-check": "tsc --noEmit",
+    "lint": "eslint --cache --cache-location /tmp/eslintcache/ 'src/**/*.ts' 'test/**/*.ts'",
+    "check-formatting": "prettier --check \"**/*.ts\"",
+    "fix-formatting": "prettier --write \"**/*.ts\"",
+    "build": "esbuild --bundle --platform=node --target=node22 --outdir=target/ src/index.ts --sourcemap --source-root=/support-service-lambdas/handlers/contributions-only-countries-api/target/",
+    "package": "pnpm type-check && pnpm lint && pnpm check-formatting && pnpm test && pnpm build && cd target && zip -qr contributions-only-countries-api.zip ./*.js.map ./*.js",
+    "cdk:test": "pnpm --filter cdk test contributions-only-countries-api",
+    "cdk:test-update": "pnpm --filter cdk test-update contributions-only-countries-api",
+    "update-lambda": "../../update-lambda.sh \"contributions-only-countries-api\"",
+    "update-stack": "../../update-stack.sh \"contributions-only-countries-api\""
+  },
+  "NOTICE1": "MANAGED FILE: to push changes see buildcheck/README.md - template: buildcheck/data/templates/handler/package.json.ts",
+  "NOTICE2": "all dependencies are defined in buildcheck/data/build.ts",
+  "dependencies": {
+    "zod": "catalog:"
+  },
+  "devDependencies": {
+    "@types/aws-lambda": "^8.10.147"
+  }
+}
