+++ handlers/contributions-only-countries-api/riff-raff.yaml
@@ -1,0 +1,26 @@
+# MANAGED FILE: to push changes see buildcheck/README.md - template: buildcheck/data/templates/handler/riff-raff.yaml.ts
+stacks:
+  - support
+regions:
+  - eu-west-1
+allowedStages:
+  - CODE
+  - PROD
+deployments:
+  contributions-only-countries-api-cloudformation:
+    type: cloud-formation
+    app: contributions-only-countries-api
+    parameters:
+      templateStagePaths:
+        CODE: contributions-only-countries-api-CODE.template.json
+        PROD: contributions-only-countries-api-PROD.template.json
+  contributions-only-countries-api:
+    type: aws-lambda
+    parameters:
+      fileName: contributions-only-countries-api.zip
+      bucketSsmLookup: true
+      prefixStack: false
+      functionNames:
+        - contributions-only-countries-api-
+    dependencies:
+      - contributions-only-countries-api-cloudformation
