--- handlers/vat-countries-list-api/README.md
@@ -1,12 +1,0 @@
-# VAT countries list API
-
-This API is an API Gateway triggered lambda to return the list of
-VAT compliant countries. It is intended to be used by SDC project
-to decide if the choice cards should appear in banners and epics (should be hidden for counties from the list).
-
-It is a public API endpoint
-
-# Urls
-
-- PROD: https://vat-countries-list-api.support.guardianapis.com/
-- CODE: https://vat-countries-list-api-code.support.guardianapis.com/
