--- handlers/vat-countries-list-api/riff-raff.yaml
@@ -1,26 +1,0 @@
-# MANAGED FILE: to push changes see buildcheck/README.md - template: buildcheck/data/templates/handler/riff-raff.yaml.ts
-stacks:
-  - support
-regions:
-  - eu-west-1
-allowedStages:
-  - CODE
-  - PROD
-deployments:
-  vat-countries-list-api-cloudformation:
-    type: cloud-formation
-    app: vat-countries-list-api
-    parameters:
-      templateStagePaths:
-        CODE: vat-countries-list-api-CODE.template.json
-        PROD: vat-countries-list-api-PROD.template.json
-  vat-countries-list-api:
-    type: aws-lambda
-    parameters:
-      fileName: vat-countries-list-api.zip
-      bucketSsmLookup: true
-      prefixStack: false
-      functionNames:
-        - vat-countries-list-api-
-    dependencies:
-      - vat-countries-list-api-cloudformation
