+++ modules/contributions-only-countries-list/index.ts
@@ -1,0 +1,1 @@
+export { countries } from './src/contributionsOnlyCountriesList';
