--- modules/vat-countries-list/index.ts
@@ -1,1 +1,0 @@
-export { countries } from './src/VATComplianceCountriesList';
